+++ src/main/java/com/runicrealms/plugin/spellapi/spells/runic/passive/Agility.java
@@ -25,7 +25,7 @@ public class Agility extends Spell {
 
     public Agility() {
         super ("Agility",
-                "Damaging an enemy has a " + PERCENT + "% chance" +
+                "Damaging an enemy has a " + (int) PERCENT + "% chance" +
                         "\nto grant you speed for " + DURATION + " second(s)!",
                 ChatColor.WHITE, ClassEnum.RUNIC, 0, 0);
         this.setIsPassive(true);
