+++ src/main/java/com/runicrealms/plugin/spellapi/spells/runic/passive/Siphon.java
@@ -23,7 +23,7 @@ public class Siphon extends Spell {
 
     public Siphon() {
         super ("Siphon",
-                "Damaging an enemy has a " + PERCENT + "% chance" +
+                "Damaging an enemy has a " + (int) PERCENT + "% chance" +
                         "\nto restore✦ " + AMOUNT + " of your mana!",
                 ChatColor.WHITE, ClassEnum.RUNIC, 0, 0);
         this.setIsPassive(true);
